Newsletter + Forms: lift Gravatar into @automattic/jetpack-components
The Gravatar component (with hovercard support, SHA-256 hashing, and a
configurable defaultImage) moves into @automattic/jetpack-components
exposed as the `./gravatar` subpath. Newsletter (subscribers stage) and
Forms (inbox + responses) both switch to it; the local copies are
removed and the per-package `js-sha256` / `@gravatar-com/hovercards`
deps move with them.

This was lifted from PR #48386 (Subscribers Dashboard) per the same
intent: one Gravatar component shared across products. The local
<Avatar> shim Newsletter relied on while #48386 wasn't on trunk goes
away too.
