+++ projects/js-packages/components/package.json
@@ -23,6 +23,11 @@
 			"types": "./build/index.d.ts",
 			"default": "./build/index.js"
 		},
+		"./gravatar": {
+			"jetpack:src": "./components/gravatar/index.tsx",
+			"types": "./build/components/gravatar/index.d.ts",
+			"default": "./build/components/gravatar/index.js"
+		},
 		"./jetpack-logo": {
 			"jetpack:src": "./components/jetpack-logo/index.tsx",
 			"types": "./build/components/jetpack-logo/index.d.ts",
@@ -53,6 +58,7 @@
 		"@automattic/jetpack-script-data": "workspace:*",
 		"@automattic/number-formatters": "workspace:*",
 		"@babel/runtime": "^7",
+		"@gravatar-com/hovercards": "0.16.0",
 		"@wordpress/admin-ui": "1.12.0",
 		"@wordpress/browserslist-config": "6.44.0",
 		"@wordpress/components": "32.6.0",
@@ -66,6 +72,7 @@
 		"@wordpress/theme": "0.11.0",
 		"@wordpress/ui": "0.11.0",
 		"clsx": "2.1.1",
+		"js-sha256": "0.11.1",
 		"prop-types": "^15.7.2",
 		"qrcode.react": "4.2.0",
 		"react-slider": "2.0.5",
