+++ projects/packages/forms/package.json
@@ -50,7 +50,6 @@
 		"@automattic/jetpack-shared-extension-utils": "workspace:*",
 		"@automattic/number-formatters": "workspace:*",
 		"@automattic/request-external-access": "1.0.1",
-		"@gravatar-com/hovercards": "0.16.0",
 		"@wordpress/admin-ui": "1.12.0",
 		"@wordpress/base-styles": "6.20.0",
 		"@wordpress/block-editor": "15.17.0",
@@ -83,7 +82,6 @@
 		"debug": "4.4.3",
 		"email-validator": "2.0.4",
 		"gridicons": "3.4.2",
-		"js-sha256": "0.11.1",
 		"libphonenumber-js": "1.12.42",
 		"lodash": "4.18.1",
 		"photon": "4.1.1",
